backend/requirements.txt
locust==2.43.2
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
alembic==1.18.3
pymysql==1.1.2
passlib==1.7.4
python-jose==3.5.0
inflection==0.5.1
bcrypt==4.0.1
pydantic-settings==2.12.0
typer==0.21.1
cryptography==46.0.4